requirements.txt
Django==1.10
psycopg2==2.6.1
django-debug-toolbar==1.4
django-widget-tweaks==1.4.1
djangorestframework==3.4.4
django-filter==0.14.0
drf-nested-routers==0.11.1
django_debug_toolbar==1.5
djangorestframework-filters==0.8.0